Feat: PR 수정사항 변경 및 function util 디렉토리 아래로 이동 [#20]
